Renaming and implement CacheWarmerInterface
